Estudos de boxing, data e BigDecimal

+++ Canco/src/com/dgama/banco/modelo/pagamento/DocumentoPagavel.java
@@ -1,10 +1,11 @@
 package com.dgama.banco.modelo.pagamento;
 
+import java.math.BigDecimal;
 import java.sql.SQLOutput;
 
 public interface DocumentoPagavel {
 
-    double getValorTotal();
+    BigDecimal getValorTotal();
     boolean estaPago();
     void quitarPagamento();
